move 'ContenedorNavegacion' into a new component
